+++ src/utils/validate.js
@@ -96,3 +96,94 @@ export const validateAddTask = (addTaskData,setErrors) => {
   setErrors(errors);
   return isVaild;
 };
+
+
+
+//neww
+
+
+// export const checkPasswordRequirements = (inputPwd) => {
+//   const requirements = [
+//       {
+//           key: "upperCaseError",
+//           regex: /[A-Z]/,
+//           message: "At least one uppercase letter is required"
+//       },
+//       {
+//           key: "lowerCaseError",
+//           regex: /[a-z]/,
+//           message: "At least one lowercase letter is required"
+//       },
+//       {
+//           key: "lengthError",
+//           regex: /^.{8,}$/, // At least 8 characters
+//           message: "Password must be at least 8 characters long"
+//       }
+//       // Add more criteria as needed
+//   ];
+
+//   // Check each requirement and collect errors
+//   const errors = requirements.reduce((acc, requirement) => {
+//       acc[requirement.key] = !requirement.regex.test(inputPwd);
+//       return acc;
+//   }, {});
+
+//   // Return errors object
+//   return {
+//       error: Object.values(errors).some((error) => error),
+//       content: requirements.filter((req) => errors[req.key]).map((req) => req.message).join(", "),
+//       // You might want to adjust the 'content' message based on your UI or error handling needs
+//       criteria: errors
+//   };
+// };
+
+const content = [
+  {
+    error: false,
+    content: 'At least one special character is required',
+    key: 'specialCharError'
+  },
+  {
+    error: false,
+    content: 'At least one lowercase letter is required',
+    key: 'lowercaseError'
+  },
+  {
+    error: false,
+    content: 'At least one number is required',
+    key: 'numberError'
+  },
+  {
+    error: false,
+    content: 'Password should be at least 8 characters long',
+    key: 'lengthError'
+  }
+];
+
+export const trackPwdRequirement = (password) => {
+  let specialCharRegex = /[!@#$%^&*()_+{}\[\]:;<>,.?~]/;
+  let lowercaseRegex = /[a-z]/;
+  let numberRegex = /\d/;
+  
+  let updatedContent = content.map((item) => ({ ...item }));
+  updatedContent.forEach((item) => {
+    switch (item.key) {
+      case 'specialCharError':
+        item.error = !specialCharRegex.test(password);
+        break;
+      case 'lowercaseError':
+        item.error = !lowercaseRegex.test(password);
+        break;
+      case 'numberError':
+        item.error = !numberRegex.test(password);
+        break;
+      case 'lengthError':
+        item.error = password.length < 8;
+        break;
+      default:
+        break;
+    }
+  });
+
+  return updatedContent;
+};
